update rent and clean
